requirements.txt
kafka-python==2.3.0
Flask==3.0.3
